+++ pnpr/.fixtures/packages/create-touch-file-one-bin/1.0.0/cli.js
@@ -1,4 +1,4 @@
 'use strict'
 const fs = require('fs')
 
-fs.writeFileSync('touch.txt', 'hello world', 'utf8')
+fs.writeFileSync('touch.txt', JSON.stringify(process.argv.slice(2)), 'utf8')
